Separate object updates in two stages, initialized and not initialized
We had some cases in which two detections would be close to each other,
and one of them would update an already established tracked object and
the other, left unmatched, would create a new tracked objects.

Then one of those detections would dissapear and the remaining detection
could be assigned to the not yet initialized tracked object instead of
the initialized one, effectively killing the initialized object, with
the not yet initialized one taking its place.

This is not ideal as it would split what should've been a single object
into two objects, creating an unnecessary id change, which is
effectively an error in tracking.

By separating tracking in two stages, and giving initialized objects
first bids this problem is avoided.