Add override argument to output_fps in Video class
When saving video from a webcam to a file, the external process can
introduce lag to each frame and make output fps be different from input
fps. So we add the ability for the user to chose output_fps manually to
avoid endung up with a video which looks sped up.