Force points being detected for the first time to have velocity = 0
This is needed because some detectors (like OpenPose) set points with
low confidence to coordinates (0, 0). And when they then get their first
real detection this creates a huge velocity vector in our KalmanFilter
and causes the tracker to start with wildly inaccurate estimations which
eventually coverge to the real detections.

This really mitigates the problem but seems to not fix it 100%. Not sure
why yet still.