Create additional functionalities to motmetrics4norfair
The following classes had been created:
-video_from_frames: to make an output video from a folder with frames, and drawing your predictions and detections
-text_file: to generate a txt file with your predictions (in the MOTChallenge format)
-Det_from_file: To process detections (given as in MOTChallenge format) from a txt file
-accumulators: This saves all your predictions informations to finally compute metrics comparing them with ground truth files (taken from a
	txt file, in MOTChallenge format)

Now motmetrics4norfair always saves your final metrics results in a txt file, and has the option to display them on terminal after being computed.
Also has the option to create a txt with predictions, and to make an output video with your predictions and detections.
Gives the option to process every folder on a path (not needing to specify each folder you want to process), and make exceptions (specify which folders
you do not want to process).